adds checks for all the other controllers

+++ controllers/goal_logs_controller.py
@@ -7,6 +7,7 @@ from util.validate_uuid4 import validate_uuid4
 from models.categories import Categories
 from models.goal_logs import GoalLogs, goal_log_schema, goal_logs_schema
 from models.goals import Goals
+from models.users import Users
 
 
 @authenticate_return_auth
@@ -73,8 +74,11 @@ def goal_log_get_by_id(req, goal_log_id):
         return jsonify({"message": "goal log found", "goal_log": goal_log_schema.dump(goal_log_query)}), 200
 
 
-@authenticate
-def goal_logs_get_all(req):
+@authenticate_return_auth
+def goal_logs_get_all(req, auth_info):
+    if auth_info.user.role != "super-admin":
+        return jsonify({"message": "unauthorized"}), 403
+
     goal_logs_query = db.session.query(GoalLogs).all()
 
     return jsonify({"message": "goal logs found", "goal_logs": goal_logs_schema.dump(goal_logs_query)}), 200
@@ -87,8 +91,8 @@ def goal_logs_get_from_auth_info(req, auth_info):
     return jsonify({"message": "goal logs found", "goal_logs": goal_logs_schema.dump(goal_logs_query)}), 200
 
 
-@authenticate
-def goal_log_update_by_id(req, goal_log_id):
+@authenticate_return_auth
+def goal_log_update_by_id(req, goal_log_id, auth_info):
     if not validate_uuid4(goal_log_id):
         return jsonify({"message": "invalid goal log id"}), 400
 
@@ -99,9 +103,29 @@ def goal_log_update_by_id(req, goal_log_id):
     if not goal_log_query:
         return jsonify({"message": "goal log not found"}), 404
 
+    if goal_log_query.user_id != auth_info.user_id and auth_info.user.role != "super-admin":
+        return jsonify({"message": "unauthorized"}), 403
+
     category_ids = post_data.get("category_ids")
     if category_ids:
         post_data.pop("category_ids")
+
+    user_id = post_data.get("user_id")
+    if "user_id" in post_data:
+        if user_id != str(auth_info.user_id) and auth_info.user.role != "super-admin":
+            return jsonify({"message": "cannot change user id"}), 400
+
+        if not validate_uuid4(user_id):
+            return jsonify({"message": "invalid user id"}), 400
+
+        user_query = db.session.query(Users).filter(Users.user_id == user_id).first()
+
+        if not user_query:
+            return jsonify({"message": "user not found"}), 404
+
+    goal_id = post_data.get("goal_id")
+    if "goal_id" in post_data:
+        return jsonify({"message": "cannot change goal id"}), 403
 
     populate_object(goal_log_query, post_data)
 
@@ -127,8 +151,8 @@ def goal_log_update_by_id(req, goal_log_id):
     return jsonify({"message": "goal log updated", "goal_log": goal_log_schema.dump(goal_log_query)}), 200
 
 
-@authenticate
-def goal_log_delete_by_id(req, goal_log_id):
+@authenticate_return_auth
+def goal_log_delete_by_id(req, goal_log_id, auth_info):
     if not validate_uuid4(goal_log_id):
         return jsonify({"message": "invalid goal log id"}), 400
 
@@ -136,6 +160,9 @@ def goal_log_delete_by_id(req, goal_log_id):
 
     if not goal_log_query:
         return jsonify({"message": "goal log not found"}), 404
+
+    if goal_log_query.user_id != str(auth_info.user_id) and auth_info.user.role != "super-admin":
+        return jsonify({"message": "unauthorized"}), 403
 
     db.session.delete(goal_log_query)
     db.session.commit()
